+++ changelog.md
@@ -17,3 +17,4 @@
 - Added search command
 - Removed name command
 - Renamed `cache_size` config option to `audio_cache_size` and increased default value to 128
+- Added translations with fallback